Creating Independent get_data file to make API calls
